% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exporter.R
\name{console_exporter}
\alias{console_exporter}
\title{Console Exporter}
\usage{
console_exporter(verbose = TRUE)
}
\arguments{
\item{verbose}{If \code{TRUE}, print detailed span information.}
}
\value{
An S3 \code{securetrace_exporter} object.
}
\description{
Creates an exporter that prints trace summaries to the console.
}
