% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{current_resource}
\alias{current_resource}
\title{Get the Current Resource}
\usage{
current_resource()
}
\value{
The current \code{securetrace_resource}, or \code{NULL} if none set.
}
\description{
Get the Current Resource
}
