% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{current_trace}
\alias{current_trace}
\title{Get the Current Active Trace}
\usage{
current_trace()
}
\value{
The active \code{Trace} object, or \code{NULL} if none.
}
\description{
Get the Current Active Trace
}
