% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exporter.R
\name{export_trace}
\alias{export_trace}
\title{Export a Trace}
\usage{
export_trace(exporter, trace)
}
\arguments{
\item{exporter}{An S3 \code{securetrace_exporter} object.}

\item{trace}{A \code{Trace} object.}
}
\value{
Invisible \code{NULL}.
}
\description{
Calls the exporter's export function with the serialized trace.
}
