% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exporter.R
\name{exporter}
\alias{exporter}
\alias{new_exporter}
\title{Create a New Exporter}
\usage{
exporter(export_fn)

new_exporter(...)
}
\arguments{
\item{export_fn}{A function that accepts a trace list (from \code{trace$to_list()}).}

\item{...}{Arguments passed to \code{\link[=exporter]{exporter()}}.}
}
\value{
An S7 object of class \code{securetrace_exporter}.
}
\description{
Factory function for creating exporter objects.
}
\examples{
# Create an exporter that counts traces
counter <- new.env(parent = emptyenv())
counter$n <- 0L
exp <- exporter(function(trace_list) {
  counter$n <- counter$n + 1L
})
}
