% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exporter.R
\name{exporter_console}
\alias{exporter_console}
\alias{console_exporter}
\title{Console Exporter}
\usage{
exporter_console(verbose = TRUE)

console_exporter(...)
}
\arguments{
\item{verbose}{If \code{TRUE}, print detailed span information.}

\item{...}{Arguments passed to \code{\link[=exporter_console]{exporter_console()}}.}
}
\value{
An S7 \code{securetrace_exporter} object.
}
\description{
Creates an exporter that prints trace summaries to the console.
}
\examples{
exp <- exporter_console(verbose = TRUE)

tr <- Trace$new("demo-run")
tr$start()
span <- Span$new("step1", type = "custom")
span$start()
span$end()
tr$add_span(span)
tr$end()
export_trace(exp, tr)
}
