% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otlp.R
\name{flush_otlp}
\alias{flush_otlp}
\title{Flush Buffered OTLP Traces}
\usage{
flush_otlp(exporter)
}
\arguments{
\item{exporter}{An OTLP exporter created by \code{\link[=exporter_otlp]{exporter_otlp()}}.}
}
\value{
Invisible \code{NULL}.
}
\description{
Forces immediate sending of any traces buffered in an OTLP exporter.
}
\examples{
\dontrun{
exp <- exporter_otlp(batch_size = 50L)
# ... export some traces ...
flush_otlp(exp)
}
}
