% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{is_trace_event}
\alias{is_trace_event}
\title{Test if an Object is a Trace Event}
\usage{
is_trace_event(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
Logical scalar.
}
\description{
Test if an Object is a Trace Event
}
