% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exporter.R
\name{new_exporter}
\alias{new_exporter}
\title{Create a New Exporter}
\usage{
new_exporter(export_fn)
}
\arguments{
\item{export_fn}{A function that accepts a trace list (from \code{trace$to_list()}).}
}
\value{
An S7 object of class \code{securetrace_exporter}.
}
\description{
Factory function for creating exporter objects.
}
\examples{
# Create an exporter that counts traces
counter <- new.env(parent = emptyenv())
counter$n <- 0L
exp <- new_exporter(function(trace_list) {
  counter$n <- counter$n + 1L
})
}
