% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{sampler_always_on}
\alias{sampler_always_on}
\title{Always-On Sampler}
\usage{
sampler_always_on()
}
\value{
A \code{securetrace_sampler} that always returns \code{TRUE}.
}
\description{
Records every trace. This is the default.
}
\examples{
s <- sampler_always_on()
s@should_sample("test", list())
}
