% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{sampler_probability}
\alias{sampler_probability}
\title{Probability Sampler}
\usage{
sampler_probability(rate = 1)
}
\arguments{
\item{rate}{Sampling rate between 0 and 1. Default 1.0 (all traces).}
}
\value{
A \code{securetrace_sampler}.
}
\description{
Records traces with a given probability.
}
\examples{
# Record 10\% of traces
s <- sampler_probability(0.1)
}
