% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{sampler_rate_limiting}
\alias{sampler_rate_limiting}
\title{Rate-Limiting Sampler}
\usage{
sampler_rate_limiting(max_per_second = 10)
}
\arguments{
\item{max_per_second}{Maximum traces per second. Default 10.}
}
\value{
A \code{securetrace_sampler}.
}
\description{
Records at most N traces per second.
}
\examples{
s <- sampler_rate_limiting(5)
}
