% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{securetrace_sampler}
\alias{securetrace_sampler}
\title{S7 class: securetrace_sampler}
\usage{
securetrace_sampler(should_sample = function() NULL)
}
\arguments{
\item{should_sample}{A function taking \code{name} and \code{metadata} arguments,
returning \code{TRUE} to record or \code{FALSE} to drop the trace.}
}
\value{
An S7 object of class \code{securetrace_sampler}.
}
\description{
S7 class: securetrace_sampler
}
