% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{set_default_exporter}
\alias{set_default_exporter}
\title{Set the Default Exporter}
\usage{
set_default_exporter(exporter)
}
\arguments{
\item{exporter}{An S3 \code{securetrace_exporter} object.}
}
\value{
Invisible \code{NULL}.
}
\description{
Set the Default Exporter
}
