% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{set_default_sampler}
\alias{set_default_sampler}
\title{Set the Default Sampler}
\usage{
set_default_sampler(sampler)
}
\arguments{
\item{sampler}{A \code{securetrace_sampler} object.}
}
\value{
Invisible \code{NULL}.
}
\description{
Set the Default Sampler
}
\examples{
# Only record 50\% of traces
set_default_sampler(sampler_probability(0.5))

# Reset to record everything
set_default_sampler(sampler_always_on())
}
