% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{set_resource}
\alias{set_resource}
\title{Set the Default Resource}
\usage{
set_resource(res)
}
\arguments{
\item{res}{A \code{securetrace_resource} object from \code{\link[=resource]{resource()}}.}
}
\value{
Invisible \code{NULL}.
}
\description{
Sets resource attributes that will be attached to all traces created
via \code{\link[=with_trace]{with_trace()}}.
}
\examples{
set_resource(resource("my-agent", service_version = "1.0.0"))

# Traces now include resource attributes
with_trace("test", {
  tr <- current_trace()
  tr$resource
})
}
