% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost.R
\name{calculate_cost}
\alias{calculate_cost}
\title{Calculate Cost for a Model Call}
\usage{
calculate_cost(model, input_tokens, output_tokens)
}
\arguments{
\item{model}{Model name string.}

\item{input_tokens}{Number of input tokens.}

\item{output_tokens}{Number of output tokens.}
}
\value{
Numeric cost in USD.
}
\description{
Calculate Cost for a Model Call
}
