% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exporter.R
\name{multi_exporter}
\alias{multi_exporter}
\title{Multi-Exporter}
\usage{
multi_exporter(...)
}
\arguments{
\item{...}{Exporter objects to combine.}
}
\value{
An S3 \code{securetrace_exporter} object.
}
\description{
Combines multiple exporters into one. When a trace is exported,
it is sent to all contained exporters.
}
