% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exporter.R
\name{new_exporter}
\alias{new_exporter}
\title{Create a New Exporter}
\usage{
new_exporter(export_fn)
}
\arguments{
\item{export_fn}{A function that accepts a trace list (from \code{trace$to_list()}).}
}
\value{
An S3 object of class \code{securetrace_exporter}.
}
\description{
Wraps an export function as an S3 exporter object.
}
