% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{trace_event}
\alias{trace_event}
\title{Create a Trace Event}
\usage{
trace_event(name, data = list(), timestamp = Sys.time())
}
\arguments{
\item{name}{Name of the event.}

\item{data}{Optional named list of event data.}

\item{timestamp}{Timestamp for the event. Defaults to current time.}
}
\value{
An S3 object of class \code{securetrace_event}.
}
\description{
Discrete point-in-time event within a span.
}
