% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost.R
\name{trace_total_cost}
\alias{trace_total_cost}
\title{Calculate Total Cost for a Trace}
\usage{
trace_total_cost(trace)
}
\arguments{
\item{trace}{A \code{Trace} object.}
}
\value{
Numeric total cost in USD.
}
\description{
Sums the cost of all spans in a trace that have a model recorded.
}
